Input EntitySchema ID instead of URL to be more explicit
Since this is intended to get the code solely from wikidata, we can
easily restrict it to that source and reduce the input to the minimum
needed information.

In principle, it would be possible to also remove the "E", but since the
EntitySchema ID E123 is one atomic group, we should train the user to
subdivide it further.